Added support for embedded PHP, JSP, ASP, etc.
I liked this plugin, but I hated that I couldn't use it on pages with <?php ?> or <% %> tags. It totally destroyed the CGI code layout. Wrote in a fix so it will detect those tags and not mess with them. Also added a "beautify_cgi" option, enabled by default, which will beautify the CGI code as if it were part of a <script> tag. While it likely won't be perfect, it seems to do a very good job with PHP and JSP.