Make the file extension check ignore case
Today, if the file extension is XML/Html rather than xml/html, the format logic could not be applied.  User will have to go to the '. jsbeautifyrc' file to manually add the capitalized extension. And they might be a lot of low/uppercase combinations. 

This change adds an "i" to the regexp so that case would not matter in the 'isTypeAllowed()' function. This way it will accept file extensions ignoring cases. 

Thanks!