Better error handling
Before this all errors were assumed to be "Node.js was not found in the default path", even when node actually executed fine.

This will now catch all uncaughtException in the node script and display them as error dialog.

Specifically catches JSBeautifier errors.

Displays error when run on a file that isn't identified.